fix: refactor code for cleanup and better structure
